+++ src/ascent/runtimes/expressions/ascent_expression_filters.hpp
@@ -509,11 +509,22 @@ public:
   virtual void execute();
 };
 
-class Bin: public ::flow::Filter
+class Bin : public ::flow::Filter
 {
 public:
   Bin();
   ~Bin();
+
+  virtual void declare_interface(conduit::Node &i);
+  virtual bool verify_params(const conduit::Node &params, conduit::Node &info);
+  virtual void execute();
+};
+
+class Bounds : public ::flow::Filter
+{
+public:
+  Bounds();
+  ~Bounds();
 
   virtual void declare_interface(conduit::Node &i);
   virtual bool verify_params(const conduit::Node &params, conduit::Node &info);